requirements.txt
django
django-admin-interface
djangorestframework